+++ src/fileLoader.cpp
@@ -10,11 +10,14 @@ using namespace std;
 using namespace filesystem;
 
 
-const string configPaths[] = 
+const string configPaths[] =
 {
-    FileLoader::GetEnvVar("XDG_CONFIG_HOME") + "/config.jsonc",
-    FileLoader::GetEnvVar("XDG_CONFIG_HOME") + "/config.json",
-    FileLoader::GetEnvVar("XDG_CONFIG_HOME") + "/config",
+    (FileLoader::GetEnvVar("XDG_CONFIG_HOME").empty() ? "" :
+     FileLoader::GetEnvVar("XDG_CONFIG_HOME") + "/") + "config.jsonc",
+    (FileLoader::GetEnvVar("XDG_CONFIG_HOME").empty() ? "" :
+     FileLoader::GetEnvVar("XDG_CONFIG_HOME") + "/") + "config.json",
+    (FileLoader::GetEnvVar("XDG_CONFIG_HOME").empty() ? "" :
+    FileLoader::GetEnvVar("XDG_CONFIG_HOME") + "/") + "config",
     "~/.config/glitch-effect/config.jsonc",
     "~/.config/glitch-effect/config.json",
     "~/.config/glitch-effect/config",