+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # vX.X.X
 * Slight optimizations
 * New pass: Vertical offset
+* Fix a bug where file loader would search for config in incorrect paths if `XDG_CONFIG_HOME` variable was empty